src/main/java/org/example/StringUtils.java
package org.example;

public class StringUtils {
    public String reverse(String input) {
        if (input == null) {
            throw new IllegalArgumentException("Input cannot be null");
        }
        return new StringBuilder(input).reverse().toString();
    }

    public boolean isPalindrome(String input) {
        if (input == null) {
            throw new IllegalArgumentException("Input cannot be null");
        }
        String reversed = reverse(input);
        return input.equals(reversed);
    }
}
